+++ spinetoolbox/config.py
@@ -21,7 +21,7 @@ import os
 from pathlib import Path
 
 # NOTE: All required Python package versions are in setup.cfg
-LATEST_PROJECT_VERSION = 9
+LATEST_PROJECT_VERSION = 10
 
 # For the Add/Update SpineOpt wizard
 REQUIRED_SPINE_OPT_VERSION = "0.6.9"